# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anatomiya",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/student": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "All Students",
                "description": "Hamma studentlarni korish",
                "operationId": "student_index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "new Student add",
                "description": "Yangi student qoshish",
                "operationId": "student_store",
                "requestBody": {
                    "description": "Student Save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password",
                                    "balance",
                                    "birthday"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Harley"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "admin123"
                                    },
                                    "balance": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "6275"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "1987-08-29"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/student/{id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Show Student",
                "description": "bitta studentni hamma malumotlarini ko'rsatadi",
                "operationId": "student_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Student"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "student_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "student id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Student Save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password",
                                    "status",
                                    "balance",
                                    "birthday"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Harley"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "admin123"
                                    },
                                    "status": {
                                        "type": "enum",
                                        "format": "text",
                                        "example": "active"
                                    },
                                    "balance": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "6275"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "1987-08-29"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Student"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "student_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "text",
                            "example": "Not Found"
                        }
                    }
                },
                "type": "object"
            },
...
            "Student": {
                "title": "Student",
                "required": [
                    "name",
                    "surname",
                    "email",
                    "status",
                    "is_blocked",
                    "password",
                    "balance"
                ],
                "properties": {
                    "name_uz": {
                        "type": "string"
                    },
                    "name_ru": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "parent_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "student": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer_token": {
            "type": "apiKey",
            "description": "Enter token in format (Bearer <token>)",
            "name": "Authorization",
            "in": "header"
        }
    }
}